Fix various issues with links
The cache_directory needs a trailing slash such that the atom feed can
correctly create it's id, which will be used to generate the link from
the HTML to the atom feed.

Also fix the XSLT because the meta and link elements in HTML5 aren't
closed so you need to generate them as text, which means that the
XSLT expressions within don't work and you need to use xsl:value-of
elements instead.