Simplify OPML again
Having to repeat the title attribute as the text attribute is
annoying. Adding type="rss" is confusing: what if the feed is an Atom
feed? All of this points to the fact that those attributes should not
be set in the first place.

Or perhaps they should be set at a later date?